Move logic to new CSharpDiagnosticAnalyzerQuickInfoProvider
